requirements.txt
influxdb-client
coc.py